chore: rename prune to purge
