+++ pkg/views/purge/server_stopped.go
@@ -1,7 +1,7 @@
 // Copyright 2024 Daytona Platforms Inc.
 // SPDX-License-Identifier: Apache-2.0
 
-package prune
+package purge
 
 import (
 	"log"
